fonts: remove some dead code
