+++ src/font/mod.rs
@@ -1,38 +1,22 @@
-use failure::{bail, format_err, Error, Fallible};
-use log::{debug, error};
-mod ftfont;
+use failure::{format_err, Error, Fallible};
 mod hbwrap;
-use self::hbwrap as harfbuzz;
 
 use std::cell::RefCell;
 use std::collections::HashMap;
 use std::rc::Rc;
 
+pub mod ftwrap;
 pub mod loader;
 pub mod rasterizer;
 pub mod shaper;
 
-pub mod system;
-pub use self::system::*;
-
-pub mod ftwrap;
-
 #[cfg(all(unix, any(feature = "fontconfig", not(target_os = "macos"))))]
 pub mod fcwrap;
 
-#[cfg(target_os = "macos")]
-pub mod coretext;
-
-#[cfg(any(target_os = "macos", windows))]
-pub mod fontloader;
-#[cfg(any(target_os = "macos", windows))]
-pub mod fontloader_and_freetype;
-
-#[cfg(any(target_os = "macos", windows))]
-pub mod fontkit;
-
 use crate::font::loader::{FontLocator, FontLocatorSelection};
+pub use crate::font::rasterizer::RasterizedGlyph;
 use crate::font::rasterizer::{FontRasterizer, FontRasterizerSelection};
+pub use crate::font::shaper::{FallbackIdx, FontMetrics, GlyphInfo};
 use crate::font::shaper::{FontShaper, FontShaperSelection};
 
 use super::config::{configuration, ConfigHandle, TextStyle};
@@ -207,156 +191,4 @@ impl FontConfiguration {
         }
         &config.font
     }
-}
-
-#[allow(dead_code)]
-pub fn shape_with_harfbuzz(
-    font: &mut dyn NamedFont,
-    font_idx: system::FallbackIdx,
-    s: &str,
-) -> Result<Vec<GlyphInfo>, Error> {
-    let features = vec![
-        // kerning
-        harfbuzz::feature_from_string("kern")?,
-        // ligatures
-        harfbuzz::feature_from_string("liga")?,
-        // contextual ligatures
-        harfbuzz::feature_from_string("clig")?,
-    ];
-
-    let mut buf = harfbuzz::Buffer::new()?;
-    buf.set_script(harfbuzz::hb_script_t::HB_SCRIPT_LATIN);
-    buf.set_direction(harfbuzz::hb_direction_t::HB_DIRECTION_LTR);
-    buf.set_language(harfbuzz::language_from_string("en")?);
-    buf.add_str(s);
-
-    {
-        let fallback = font.get_fallback(font_idx).map_err(|e| {
-            let chars: Vec<u32> = s.chars().map(|c| c as u32).collect();
-            e.context(format!("while shaping {:x?}", chars))
-        })?;
-        fallback.harfbuzz_shape(&mut buf, Some(features.as_slice()));
-    }
-
-    let infos = buf.glyph_infos();
-    let positions = buf.glyph_positions();
-
-    let mut cluster = Vec::new();
-
-    let mut last_text_pos = None;
-    let mut first_fallback_pos = None;
-
-    // Compute the lengths of the text clusters.
-    // Ligatures and combining characters mean
-    // that a single glyph can take the place of
-    // multiple characters.  The 'cluster' member
-    // of the glyph info is set to the position
-    // in the input utf8 text, so we make a pass
-    // over the set of clusters to look for differences
-    // greater than 1 and backfill the length of
-    // the corresponding text fragment.  We need
-    // the fragments to properly handle fallback,
-    // and they're handy to have for debugging
-    // purposes too.
-    let mut sizes = Vec::with_capacity(s.len());
-    for (i, info) in infos.iter().enumerate() {
-        let pos = info.cluster as usize;
-        let mut size = 1;
-        if let Some(last_pos) = last_text_pos {
-            let diff = pos - last_pos;
-            if diff > 1 {
-                sizes[i - 1] = diff;
-            }
-        } else if pos != 0 {
-            size = pos;
-        }
-        last_text_pos = Some(pos);
-        sizes.push(size);
-    }
-    if let Some(last_pos) = last_text_pos {
-        let diff = s.len() - last_pos;
-        if diff > 1 {
-            let last = sizes.len() - 1;
-            sizes[last] = diff;
-        }
-    }
-    //debug!("sizes: {:?}", sizes);
-
-    // Now make a second pass to determine if we need
-    // to perform fallback to a later font.
-    // We can determine this by looking at the codepoint.
-    for (i, info) in infos.iter().enumerate() {
-        let pos = info.cluster as usize;
-        if info.codepoint == 0 {
-            if first_fallback_pos.is_none() {
-                // Start of a run that needs fallback
-                first_fallback_pos = Some(pos);
-            }
-        } else if let Some(start_pos) = first_fallback_pos {
-            // End of a fallback run
-            //debug!("range: {:?}-{:?} needs fallback", start, pos);
-
-            let substr = &s[start_pos..pos];
-            let mut shape = match shape_with_harfbuzz(font, font_idx + 1, substr) {
-                Ok(shape) => Ok(shape),
-                Err(e) => {
-                    error!("{:?} for {:?}", e, substr);
-                    if font_idx == 0 && s == "?" {
-                        bail!("unable to find any usable glyphs for `?` in font_idx 0");
-                    }
-                    shape_with_harfbuzz(font, 0, "?")
-                }
-            }?;
-
-            // Fixup the cluster member to match our current offset
-            for mut info in &mut shape {
-                info.cluster += start_pos as u32;
-            }
-            cluster.append(&mut shape);
-
-            first_fallback_pos = None;
-        }
-        if info.codepoint != 0 {
-            if s.is_char_boundary(pos) && s.is_char_boundary(pos + sizes[i]) {
-                let text = &s[pos..pos + sizes[i]];
-                //debug!("glyph from `{}`", text);
-                cluster.push(GlyphInfo::new(text, font_idx, info, &positions[i]));
-            } else {
-                cluster.append(&mut shape_with_harfbuzz(font, 0, "?")?);
-            }
-        }
-    }
-
-    // Check to see if we started and didn't finish a
-    // fallback run.
-    if let Some(start_pos) = first_fallback_pos {
-        let substr = &s[start_pos..];
-        if false {
-            debug!(
-                "at end {:?}-{:?} needs fallback {}",
-                start_pos,
-                s.len() - 1,
-                substr,
-            );
-        }
-        let mut shape = match shape_with_harfbuzz(font, font_idx + 1, substr) {
-            Ok(shape) => Ok(shape),
-            Err(e) => {
-                error!("{:?} for {:?}", e, substr);
-                if font_idx == 0 && s == "?" {
-                    bail!("unable to find any usable glyphs for `?` in font_idx 0");
-                }
-                shape_with_harfbuzz(font, 0, "?")
-            }
-        }?;
-        // Fixup the cluster member to match our current offset
-        for mut info in &mut shape {
-            info.cluster += start_pos as u32;
-        }
-        cluster.append(&mut shape);
-    }
-
-    //debug!("shaped: {:#?}", cluster);
-
-    Ok(cluster)
 }
